+++ NEWS.md
@@ -41,6 +41,7 @@ New library features
 Standard library changes
 ------------------------
 
+* `range` accepts either `stop` or `length` as a sole keyword argument ([#39241])
 * The `length` function on certain ranges of certain specific element types no longer checks for integer
   overflow in most cases. The new function `checked_length` is now available, which will try to use checked
   arithmetic to error if the result may be wrapping. Or use a package such as SaferIntegers.jl when